log an error when we are unable to read/write an API key to the netrc file due to permissions

+++ wandb/sdk/lib/apikey.py
@@ -1,10 +1,13 @@
 """apikey util."""
 
+from __future__ import annotations
+
 import os
 import platform
 import stat
 import sys
 import textwrap
+from enum import Enum
 from functools import partial
 
 # import Literal
@@ -32,7 +35,15 @@ LOGIN_CHOICES = [
     LOGIN_CHOICE_DRYRUN,
 ]
 
+
+class _NetrcPermissions(Enum):
+    NETRC_EXISTS = os.F_OK
+    NETRC_READ_ACCESS = os.R_OK
+    NETRC_WRITE_ACCESS = os.W_OK
+
+
 Mode = Literal["allow", "must", "never", "false", "true"]
+
 
 if TYPE_CHECKING:
     from wandb.sdk.wandb_settings import Settings
@@ -169,6 +180,26 @@ def prompt_api_key(  # noqa: C901
     return key
 
 
+def check_netrc_access(
+    netrc_path: str,
+) -> dict[_NetrcPermissions, bool]:
+    """Check if we can read and write to the netrc file."""
+    try:
+        st = os.stat(netrc_path)
+        return {
+            _NetrcPermissions.NETRC_EXISTS: True,
+            _NetrcPermissions.NETRC_WRITE_ACCESS: bool(st.st_mode & stat.S_IWUSR),
+            _NetrcPermissions.NETRC_READ_ACCESS: bool(st.st_mode & stat.S_IRUSR),
+        }
+    except FileNotFoundError:
+        # If the netrc file doesn't exist, we will create it.
+        return {
+            _NetrcPermissions.NETRC_EXISTS: False,
+            _NetrcPermissions.NETRC_WRITE_ACCESS: True,
+            _NetrcPermissions.NETRC_READ_ACCESS: True,
+        }
+
+
 def write_netrc(host: str, entity: str, key: str) -> Optional[bool]:
     """Add our host and key to .netrc."""
     _, key_suffix = key.split("-", 1) if "-" in key else ("", key)
@@ -178,20 +209,34 @@ def write_netrc(host: str, entity: str, key: str) -> Optional[bool]:
                 key_suffix, len(key_suffix)
             )
         )
-        return None
+        return False
+
+    normalized_host = urlparse(host).netloc.split(":")[0]
+    netrc_path = get_netrc_file_path()
+    netrc_access = check_netrc_access(netrc_path)
+
+    if (
+        netrc_access[_NetrcPermissions.NETRC_EXISTS]
+        and not netrc_access[_NetrcPermissions.NETRC_WRITE_ACCESS]
+    ):
+        wandb.termwarn(f"You do not have write permissions for {netrc_path}")
+        wandb.termwarn("We will be unable to save/update your API key.")
+        return False
+    if (
+        netrc_access[_NetrcPermissions.NETRC_EXISTS]
+        and not netrc_access[_NetrcPermissions.NETRC_READ_ACCESS]
+    ):
+        wandb.termwarn(f"You do not have read permissions for {netrc_path}")
+        wandb.termwarn("We will be unable to save/update your API key.")
+        return False
+
     try:
-        normalized_host = urlparse(host).netloc.split(":")[0]
-        netrc_path = get_netrc_file_path()
-        wandb.termlog(
-            f"Appending key for {normalized_host} to your netrc file: {netrc_path}"
-        )
         machine_line = f"machine {normalized_host}"
         orig_lines = None
-        try:
+        if netrc_access[_NetrcPermissions.NETRC_EXISTS]:
             with open(netrc_path) as f:
                 orig_lines = f.read().strip().split("\n")
-        except OSError:
-            pass
+
         with open(netrc_path, "w") as f:
             if orig_lines:
                 # delete this machine from the file if it's already there.
@@ -205,20 +250,24 @@ def write_netrc(host: str, entity: str, key: str) -> Optional[bool]:
                         skip -= 1
                     else:
                         f.write("{}\n".format(line))
+
+            wandb.termlog(
+                f"Appending key for {normalized_host} to your netrc file: {netrc_path}"
+            )
             f.write(
                 textwrap.dedent(
                     """\
-            machine {host}
-              login {entity}
-              password {key}
-            """
+                    machine {host}
+                      login {entity}
+                      password {key}
+                    """
                 ).format(host=normalized_host, entity=entity, key=key)
             )
         os.chmod(netrc_path, stat.S_IRUSR | stat.S_IWUSR)
         return True
     except OSError:
-        wandb.termerror(f"Unable to read {netrc_path}")
-        return None
+        wandb.termerror(f"Unable to write {netrc_path}")
+        return False
 
 
 def write_key(
@@ -249,7 +298,19 @@ def api_key(settings: Optional["Settings"] = None) -> Optional[str]:
         settings = wandb.setup().settings
     if settings.api_key:
         return settings.api_key
-    auth = get_netrc_auth(settings.base_url)
-    if auth:
-        return auth[-1]
+
+    netrc_access = check_netrc_access(get_netrc_file_path())
+
+    if netrc_access[_NetrcPermissions.NETRC_EXISTS]:
+        auth = get_netrc_auth(settings.base_url)
+        if auth:
+            return auth[-1]
+
+    if (
+        netrc_access[_NetrcPermissions.NETRC_EXISTS]
+        and not netrc_access[_NetrcPermissions.NETRC_READ_ACCESS]
+    ):
+        wandb.termwarn(f"You do not have read permissions for {get_netrc_file_path()}")
+        wandb.termwarn("You will be prompted for your API key.")
+
     return None
